requirements.txt
python-telegram-bot
mcp
fastmcp
python-docx
google-generativeai
python-dotenv
